+++ kontiki_tui/components/tabs.py
@@ -2,6 +2,7 @@ from textual.widgets import Static, TabbedContent
 
 from kontiki_tui.components.events import EventsTab
 from kontiki_tui.components.exceptions import ExceptionsTab
+from kontiki_tui.components.incidents import IncidentsTab
 from kontiki_tui.components.log import LogTab
 from kontiki_tui.components.services import ServicesTab
 from kontiki_tui.components.settings import SettingsTab
@@ -14,6 +15,7 @@ class KontikiTabs(Static):
         def __init__(self):
             super().__init__(
                 "Services",
+                "Incidents",
                 "Events",
                 "Exceptions",
                 "Logs",
@@ -27,6 +29,7 @@ class KontikiTabs(Static):
     def compose(self):
         with self.KontikiContent():
             yield ServicesTab(id_="services")
+            yield IncidentsTab(id_="incidents")
             yield EventsTab(id_="events")
             yield ExceptionsTab(id_="exceptions")
             yield LogTab(id_="log")